+++ src/wte/helpers/frontend.py
@@ -393,3 +393,48 @@ def user_admin_menubar(request, user):
                      attrs={'class': 'alert post-link',
                             'data-wte-confirm': confirm_delete('user', user.display_name, False)})
     return builder.generate()
+
+def part_change_notification_text(request, part):
+    """Generates a default notification text to be emailed when changes
+    are made to a part.
+
+    :param request: The current request to get the current user from
+    :type request: :class:`~pyramid.request.Request`
+    :param part: The part that is being changed
+    :type part: :class:`~wte.models.Part`
+    :return: The default notification text
+    :rtype: ``str``
+    """
+    if part.type == 'module':
+        return '''Hello,
+
+This is to let you know that changes have been made to the %s module.
+
+Please reload the module to see the changes.
+
+%s''' % (part.title,
+         request.current_user.display_name) 
+    elif part.type == 'part':
+        return '''Hello,
+
+This is to let you know that changes have been made to the %s %s of the %s module.
+
+Please reload the %s to see the changes.
+
+%s''' % (part.title,
+         part.label if part.label else 'part',
+         part.parent.title,
+         part.label if part.label else 'part',
+         request.current_user.display_name) 
+    elif part.type == 'page':
+        return '''Hello,
+
+This is to let you know that changes have been made to the page "%s" in the %s %s of the %s module.
+
+Please reload that page to see the changes.
+
+%s''' % (part.title,
+         part.parent.title,
+         part.parent.label if part.parent.label else 'part',
+         part.parent.parent.title,
+         request.current_user.display_name)